requirements.txt
psycopg2-binary>=2.8.6
django>=3.1
django-debug-toolbar>=3.2
Pillow>=8.1.0
django-silk>=4.1.0
asgiref~=3.3.1
django-ckeditor>=6.0.0
django-ajax-selects>=2.1.0